ai_gateway/pyproject.toml
[project]
name = "tender-ai-gateway"
version = "0.1.0"
description = "AI gateway for model routing and BYOK credential proxy."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "pydantic-settings>=2.4,<3.0",
  "uvicorn[standard]>=0.30,<1.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27,<1.0",
  "pytest>=8.3,<9.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["tender_ai_gateway"]

[tool.pytest.ini_options]
testpaths = ["tests"]
